requirements.txt
Django==1.11
djangorestframework==3.8.2
django-oauth-toolkit
django-jinja
psycopg2==2.7.4
psycopg2-binary
pytz==2018.4
requests
django-querycount
django-mptt
xlrd
xlwt
reportlab
django-redis
redis
celery
django-celery-email
django-celery-beat
django-celery-results